allow usage of CPPFLAGS, CFLAGS and LDFLAGS if exported in the environment
In this case Debian, and probably other distributions, export common
build flags to the environment. This small change picks them up, and
if not defined sticks to just define CFLAGS with "-O2".

The usage of "?=" is supported in GNU make and the OpenBSD make.
FreeBSD is handled separately anyway, so the major platforms should
be covered.